favBooks from book list

+++ src/app/pages/wishlist/wishlist.component.ts
@@ -3,7 +3,7 @@ import { Component, inject, OnInit } from '@angular/core';
 import { UserService } from '@auth-module';
 import { BookCardComponent, BooksView } from '@book-module';
 import { BackButtonComponent } from '@components';
-import { map } from 'rxjs';
+import { take, tap } from 'rxjs';
 import { FavouriteBookService } from './services/favourite-book.service';
 
 @Component({
@@ -20,17 +20,15 @@ export class WishlistComponent implements OnInit{
   userId!: number;
 
   ngOnInit(): void {
-    this.getUserData();
-     this.favouriteBookService.loadFavouriteBooks(this.userId);
-  }
-
-  getUserData() {
     this.userService.getCuurentUserSafeData().pipe(
-      map(user => this.userId = user.userId)
-    ).subscribe()
+      take(1), 
+      tap(user => this.userId = user.userId),
+      tap(user => this.favouriteBookService.loadFavouriteBooks(user.userId)),
+    ).subscribe();
   }
 
   onBookDeleteFromFavouritesEvent(book: BooksView): void {
+    if(!this.userId) return;
     this.favouriteBookService.removeBookFromFavourite(book, this.userId);
   }
